+++ photutils/psf/epsf.py
@@ -16,7 +16,7 @@ import numpy as np
 
 from .epsf_stars import EPSFStar, EPSFStars, LinkedEPSFStar
 from .models import EPSFModel
-from ..centroids import centroid_com
+from ..centroids import centroid_epsf
 from ..extern.sigma_clipping import SigmaClip
 
 try:
@@ -35,8 +35,7 @@ class EPSFFitter:
     Parameters
     ----------
     fitter : `astropy.modeling.fitting.Fitter`, optional
-        A `~astropy.modeling.fitting.Fitter` object.  The default is
-        `~astropy.modeling.fitting.LevMarLSQFitter`.
+        A `~astropy.modeling.fitting.Fitter` object.
 
     fit_boxsize : int, tuple of int, or `None`, optional
         The size (in pixels) of the box centered on the star to be used
@@ -46,8 +45,7 @@ class EPSFFitter:
         of size ``fit_boxsize`` will be used.  If ``fit_boxsize`` has
         two elements, they should be in ``(ny, nx)`` order.  The size
         must be greater than or equal to 3 pixels for both axes.  If
-        `None`, the fitter will use the entire star image.  The default
-        is 5.
+        `None`, the fitter will use the entire star image.
 
     fitter_kwargs : dict-like, optional
         Any additional keyword arguments (except ``x``, ``y``, ``z``, or
@@ -185,28 +183,24 @@ class EPSFFitter:
             x0 = 0
             y0 = 0
 
-        scaled_data = data / np.prod(epsf._oversampling)
-
-        # define positions in the ePSF oversampled grid
+        # Define positions in the undersampled grid. The fitter will
+        # evaluate on the defined interpolation grid, currently in the
+        # range [0, len(undersampled grid)].
         yy, xx = np.indices(data.shape, dtype=np.float)
-        xx = (xx - (star.cutout_center[0] - x0)) * epsf._oversampling[0]
-        yy = (yy - (star.cutout_center[1] - y0)) * epsf._oversampling[1]
+        xx = xx + x0 - star.cutout_center[0]
+        yy = yy + y0 - star.cutout_center[1]
 
         # define the initial guesses for fitted flux and shifts
         epsf.flux = star.flux
         epsf.x_0 = 0.0
         epsf.y_0 = 0.0
 
-        # create copy to avoid overwriting original oversampling factor
-        _epsf = epsf.copy()
-        _epsf._oversampling = np.array([1., 1.])
-
         try:
-            fitted_epsf = fitter(model=_epsf, x=xx, y=yy, z=scaled_data,
+            fitted_epsf = fitter(model=epsf, x=xx, y=yy, z=data,
                                  weights=weights, **fitter_kwargs)
         except TypeError:
             # fitter doesn't support weights
-            fitted_epsf = fitter(model=_epsf, x=xx, y=yy, z=scaled_data,
+            fitted_epsf = fitter(model=epsf, x=xx, y=yy, z=data,
                                  **fitter_kwargs)
 
         fit_error_status = 0
@@ -219,10 +213,8 @@ class EPSFFitter:
             fit_info = None
 
         # compute the star's fitted position
-        x_center = (star.cutout_center[0] +
-                    (fitted_epsf.x_0.value / epsf._oversampling[0]))
-        y_center = (star.cutout_center[1] +
-                    (fitted_epsf.y_0.value / epsf._oversampling[1]))
+        x_center = star.cutout_center[0] + fitted_epsf.x_0.value
+        y_center = star.cutout_center[1] + fitted_epsf.y_0.value
 
         star = copy.deepcopy(star)
         star.cutout_center = (x_center, y_center)
@@ -245,7 +237,7 @@ class EPSFBuilder:
 
     Parameters
     ----------
-    oversampling : float or tuple of two floats, optional
+    oversampling : int or tuple of two int, optional
         The oversampling factor(s) of the ePSF relative to the input
         ``stars`` along the x and y axes. The ``oversampling`` can
         either be a single float or a tuple of two floats of the form
@@ -261,73 +253,63 @@ class EPSFBuilder:
         have odd sizes along both axes to ensure a well-defined central
         pixel.
 
+    norm_radius : float, optional
+        The pixel radius over which the ePSF is normalized.
+
     smoothing_kernel : {'quartic', 'quadratic'}, 2D `~numpy.ndarray`, or `None`
         The smoothing kernel to apply to the ePSF.  The predefined
         ``'quartic'`` and ``'quadratic'`` kernels are derived from
         fourth and second degree polynomials, respectively.
         Alternatively, a custom 2D array can be input.  If `None` then
-        no smoothing will be performed.  The default is ``'quartic'``.
+        no smoothing will be performed.
 
     recentering_func : callable, optional
         A callable object (e.g. function or class) that is used to
         calculate the centroid of a 2D array.  The callable must accept
-        a 2D `~numpy.ndarray`, have a ``mask`` keyword and optionally an
-        ``error`` keyword.  The callable object must return a tuple of
-        two 1D `~numpy.ndarray`\\s, representing the x and y centroids.
-        The default is `~photutils.centroids.centroid_com`.
-
-    recentering_boxsize : float or tuple of two floats, optional
-        The size (in pixels) of the box used to calculate the centroid
-        of the ePSF during each build iteration.  If a single integer
-        number is provided, then a square box will be used.  If two
-        values are provided, then they should be in ``(ny, nx)`` order.
-        The default is 5.
+        a 2D `~numpy.ndarray`, have a ``mask`` keyword and optionally
+        ``error`` and ``oversampling`` keywords.  The callable object must return
+        a tuple of two 1D `~numpy.ndarray` variables, representing the x and y
+        centroids.
 
     recentering_maxiters : int, optional
         The maximum number of recentering iterations to perform during
-        each ePSF build iteration.  The default is 20.
+        each ePSF build iteration.
 
     fitter : `EPSFFitter` object, optional
-        A `EPSFFitter` object use to fit the ePSF to stars.  The default
-        fitter used by `EPSFFitter` is
-        `~astropy.modeling.fitting.LevMarLSQFitter`.  See the
-        `EPSFFitter` documentation its options.
-
-    center_accuracy : float, optional
-        The desired accuracy for the centers of stars.  The building
-        iterations will stop if the centers of all the stars change by
-        less than ``center_accuracy`` pixels between iterations.  All
-        stars must meet this condition for the loop to exit.  The
-        default is 1.0e-3.
+        A `EPSFFitter` object use to fit the ePSF to stars.  To set fitter
+        options, a new object with specific options should be passed in - the
+        default uses simply the default options.  To see more of these options,
+        see the `EPSFFitter` documentation.
 
     maxiters : int, optional
-        The maximum number of iterations to perform.  If the
-        ``center_accuracy`` is met, then the iterations will stop prior
-        to ``maxiters``.  The default is 10.
+        The maximum number of iterations to perform.
 
     progress_bar : bool, option
         Whether to print the progress bar during the build iterations.
-        The default is `True`.
     """
 
-    def __init__(self, oversampling=4., shape=None, smoothing_kernel='quartic',
-                 recentering_func=centroid_com, recentering_boxsize=(5, 5),
-                 recentering_maxiters=20, fitter=EPSFFitter(), center_accuracy=1.0e-3,
-                 maxiters=10, progress_bar=True):
+    def __init__(self, oversampling=4., shape=None,
+                 smoothing_kernel='quartic', norm_radius=5.5, shift_val=0.5,
+                 recentering_func=centroid_epsf, recentering_maxiters=20,
+                 fitter=EPSFFitter(), maxiters=10, progress_bar=True):
 
-        if oversampling <= 0.0:
-            raise ValueError('oversampling must be a positive number.')
-        oversampling = np.atleast_1d(oversampling).astype(float)
+        if oversampling is None:
+            raise ValueError("'oversampling' must be specified.")
+        oversampling = np.atleast_1d(oversampling).astype(int)
         if len(oversampling) == 1:
             oversampling = np.repeat(oversampling, 2)
+        if np.any(oversampling % 2 != 0):
+                raise ValueError('Oversampling factor must be a multiple of two')
+        if np.any(oversampling <= 0.0):
+            raise ValueError('oversampling must be a positive number.')
+        self._norm_radius = norm_radius
+        self._shift_val = shift_val
         self.oversampling = oversampling
         self.shape = self._init_img_params(shape)
         if self.shape is not None:
             self.shape = self.shape.astype(int)
 
         self.recentering_func = recentering_func
-        self.recentering_boxsize = self._init_img_params(recentering_boxsize)
-        self.recentering_boxsize = self.recentering_boxsize.astype(int)
         self.recentering_maxiters = recentering_maxiters
 
         self.smoothing_kernel = smoothing_kernel
@@ -336,20 +318,16 @@ class EPSFBuilder:
             raise TypeError('fitter must be an EPSFFitter instance.')
         self.fitter = fitter
 
-        if center_accuracy <= 0.0:
-            raise ValueError('center_accuracy must be a positive number.')
-        self.center_accuracy_sq = center_accuracy**2
-
         maxiters = int(maxiters)
         if maxiters <= 0:
-            raise ValueError('maxiters must be a positive number.')
+            raise ValueError("'maxiters' must be a positive number.")
         self.maxiters = maxiters
 
         self.progress_bar = progress_bar
 
         # TODO: allow custom SigmaClip object after faster SigmaClip
         # is available in astropy (>=3.1)
-        self.sigclip = SigmaClip(sigma=3., cenfunc='median', maxiters=10)
+        self.sigclip = SigmaClip(sigma=2.5, cenfunc='mean', maxiters=10)
 
         # store some data during each ePSF build iteration
         self._nfit_failed = []
@@ -400,6 +378,8 @@ class EPSFBuilder:
             The initial ePSF model.
         """
 
+        norm_radius = self._norm_radius
+        shift_val = self._shift_val
         oversampling = self.oversampling
         shape = self.shape
 
@@ -409,21 +389,27 @@ class EPSFBuilder:
             if len(shape) == 1:
                 shape = np.repeat(shape, 2)
         else:
-            x_shape = np.int(np.ceil(stars._max_shape[1] *
-                                     oversampling[0]))
-            y_shape = np.int(np.ceil(stars._max_shape[0] *
-                                     oversampling[1]))
+            # Stars class should have odd-sized dimensions, and thus we get the
+            # oversampled shape as oversampling * len + 1; if len=25, then
+            # newlen=101, for example.
+            x_shape = np.int(np.ceil(stars._max_shape[0]) * oversampling[0] + 1)
+            y_shape = np.int(np.ceil(stars._max_shape[1]) * oversampling[1] + 1)
             shape = np.array((y_shape, x_shape))
 
-        # ensure odd sizes
+        # verify odd sizes of shape
         shape = [(i + 1) if i % 2 == 0 else i for i in shape]
 
         data = np.zeros(shape, dtype=np.float)
-        xcenter = (shape[1] - 1) / 2.
-        ycenter = (shape[0] - 1) / 2.
+        # ePSF origin should be in the undersampled pixel units, not the oversampled
+        # grid units. The middle, fractional (as we wish for the center of the
+        # pixel, so the center should be at (v.5, w.5) detector pixels) value is
+        # simply the average of the two values at the extremes.
+        xcenter = stars._max_shape[0] / 2.
+        ycenter = stars._max_shape[1] / 2.
 
         epsf = EPSFModel(data=data, origin=(xcenter, ycenter),
-                         normalize=False, oversampling=oversampling)
+                         oversampling=oversampling, norm_radius=norm_radius,
+                         shift_val=shift_val)
 
         return epsf
 
@@ -453,11 +439,10 @@ class EPSFBuilder:
             image contains NaNs where there is no data.
         """
 
-        # find the integer index of EPSFStar pixels in the oversampled
-        # ePSF grid
-        x = epsf._oversampling[0] * star._xidx_centered
-        y = epsf._oversampling[1] * star._yidx_centered
-        epsf_xcenter, epsf_ycenter = epsf.origin
+        x = epsf.oversampling[0] * star._xidx_centered
+        y = epsf.oversampling[1] * star._yidx_centered
+        epsf_xcenter, epsf_ycenter = (int((epsf.data.shape[1] - 1) / 2),
+                                      int((epsf.data.shape[0] - 1) / 2))
         xidx = _py2intround(x + epsf_xcenter)
         yidx = _py2intround(y + epsf_ycenter)
 
@@ -466,15 +451,13 @@ class EPSFBuilder:
         xidx = xidx[mask]
         yidx = yidx[mask]
 
-        # Compute the normalized residual image by subtracting the
-        # normalized ePSF model from the normalized star at the location
-        # of the star in the undersampled grid.  Then, resample the
-        # normalized residual image in the oversampled ePSF grid.
-        # [(star - (epsf * xov * yov)) / (xov * yov)]
-        # --> [(star / (xov * yov)) - epsf]
-        stardata = ((star._data_values_normalized / np.prod(epsf._oversampling)) -
-                    epsf.evaluate(x=x, y=y, flux=1.0, x_0=0.0, y_0=0.0,
-                                  use_oversampling=False))
+        x = star._xidx_centered
+        y = star._yidx_centered
+
+        # Compute the normalized residual by subtracting the ePSF model from
+        # the normalized star at the location of the star in the undersampled grid.
+        stardata = (star._data_values_normalized -
+                    epsf.evaluate(x=x, y=y, flux=1.0, x_0=0.0, y_0=0.0))
 
         resampled_img = np.full(epsf.shape, np.nan)
         resampled_img[yidx, xidx] = stardata[mask]
@@ -495,16 +478,16 @@ class EPSFBuilder:
 
         Returns
         -------
-        star_imgs : 3D `~numpy.ndarray`
+        epsf_resid : 3D `~numpy.ndarray`
             A 3D cube containing the resampled residual images.
         """
 
         shape = (stars.n_good_stars, epsf.shape[0], epsf.shape[1])
-        star_imgs = np.zeros(shape)
+        epsf_resid = np.zeros(shape)
         for i, star in enumerate(stars.all_good_stars):
-            star_imgs[i, :, :] = self._resample_residual(star, epsf)
+            epsf_resid[i, :, :] = self._resample_residual(star, epsf)
 
-        return star_imgs
+        return epsf_resid
 
     def _smooth_epsf(self, epsf_data):
         """
@@ -568,17 +551,13 @@ class EPSFBuilder:
 
         return convolve(epsf_data, kernel)
 
-    def _recenter_epsf(self, epsf_data, epsf, centroid_func=centroid_com,
-                       box_size=5, maxiters=20, center_accuracy=1.0e-4):
+    def _recenter_epsf(self, epsf, centroid_func=centroid_epsf):
         """
         Calculate the center of the ePSF data and shift the data so the
         ePSF center is at the center of the ePSF data array.
 
         Parameters
         ----------
-        epsf_data : 2D `~numpy.ndarray`
-            A 2D array containing the ePSF image.
-
         epsf : `EPSFModel` object
             The ePSF model.
 
@@ -587,26 +566,8 @@ class EPSFBuilder:
             calculate the centroid of a 2D array.  The callable must
             accept a 2D `~numpy.ndarray`, have a ``mask`` keyword and
             optionally an ``error`` keyword.  The callable object must
-            return a tuple of two 1D `~numpy.ndarray`\\s, representing
-            the x and y centroids.  The default is
-            `~photutils.centroids.centroid_com`.
-
-        recentering_boxsize : float or tuple of two floats, optional
-            The size (in pixels) of the box used to calculate the
-            centroid of the ePSF during each build iteration.  If a
-            single integer number is provided, then a square box will be
-            used.  If two values are provided, then they should be in
-            ``(ny, nx)`` order.  The default is 5.
-
-        maxiters : int, optional
-            The maximum number of recentering iterations to perform.
-            The default is 20.
-
-        center_accuracy : float, optional
-            The desired accuracy for the centers of stars.  The building
-            iterations will stop if the center of the ePSF changes by
-            less than ``center_accuracy`` pixels between iterations.
-            The default is 1.0e-4.
+            return a tuple of two 1D `~numpy.ndarray` variables, representing
+            the x and y centroids.
 
         Returns
         -------
@@ -614,56 +575,39 @@ class EPSFBuilder:
             The recentered ePSF data.
         """
 
-        # Define an EPSFModel for the input data.  This EPSFModel will be
-        # used to evaluate the model on a shifted pixel grid to place the
-        # centroid at the array center.
-        epsf = EPSFModel(data=epsf_data, origin=epsf.origin, normalize=False,
-                         oversampling=epsf.oversampling)
-
-        epsf.fill_value = 0.0
         xcenter, ycenter = epsf.origin
 
-        dx_total = 0
-        dy_total = 0
-        y, x = np.indices(epsf_data.shape, dtype=np.float)
+        y, x = np.indices(epsf._data.shape, dtype=np.float)
+        x /= epsf.oversampling[0]
+        y /= epsf.oversampling[1]
 
-        iter_num = 0
-        center_accuracy_sq = center_accuracy ** 2
-        center_dist_sq = center_accuracy_sq + 1.e6
-        center_dist_sq_prev = center_dist_sq + 1
-        while (iter_num < maxiters and
-               center_dist_sq >= center_accuracy_sq):
+        mask = ~np.isfinite(epsf._data)
 
-            iter_num += 1
-
-            # extract a cutout from the ePSF
-            slices_large, _ = overlap_slices(epsf_data.shape, box_size,
-                                             (ycenter, xcenter))
-            epsf_cutout = epsf_data[slices_large]
-            mask = ~np.isfinite(epsf_cutout)
-
+        try:
             # find a new center position
-            xcenter_new, ycenter_new = centroid_func(epsf_cutout, mask=mask)
-            xcenter_new += slices_large[1].start
-            ycenter_new += slices_large[0].start
+            xcenter_new, ycenter_new = centroid_func(epsf._data, mask=mask,
+                                                     oversampling=epsf.oversampling,
+                                                     shift_val=epsf._shift_val)
+        except TypeError:
+            pass
+        try:
+            xcenter_new, ycenter_new = centroid_func(epsf._data, mask=mask,
+                                                     oversampling=epsf.oversampling)
+        # default centroid_epsf overloaded, or otherwise a function that does not accept
+        # oversampling or shift_val, in which case just pass data and mask
+        except TypeError:
+            xcenter_new, ycenter_new = centroid_func(epsf._data, mask=mask)
 
-            # calculate the shift
-            dx = xcenter - xcenter_new
-            dy = ycenter - ycenter_new
-            center_dist_sq = dx**2 + dy**2
-            if center_dist_sq >= center_dist_sq_prev:  # don't shift
-                break
-            center_dist_sq_prev = center_dist_sq
+        # Calculate the shift; dx = i - x_star so if dx was positively
+        # incremented then x_star was negatively incremented for a given i.
+        # We will therefore actually subsequently subtract dx from xcenter
+        # (or x_star).
+        dx = xcenter_new - xcenter
+        dy = ycenter_new - ycenter
 
-            # Resample the ePSF data to a shifted grid to place the
-            # centroid in the center of the central pixel.  The shift is
-            # always performed on the input epsf_data.
-            dx_total += dx  # accumulated shifts for the input epsf_data
-            dy_total += dy
-            epsf_data = epsf.evaluate(x=x, y=y, flux=1.0,
-                                      x_0=xcenter + dx_total,
-                                      y_0=ycenter + dy_total,
-                                      use_oversampling=False)
+        epsf_data = epsf.evaluate(x=x, y=y, flux=1.0,
+                                  x_0=xcenter - dx,  # subtract dx from x_0
+                                  y_0=ycenter - dy)  # even if positive
 
         return epsf_data
 
@@ -697,59 +641,55 @@ class EPSFBuilder:
             # improve the input ePSF
             epsf = copy.deepcopy(epsf)
 
-        # compute a 3D stack of 2D residual images
-        residuals = self._resample_residuals(stars, epsf)
+        for _ in range(self.recentering_maxiters):
+            # compute a 3D stack of 2D residual images
+            residuals = self._resample_residuals(stars, epsf)
+            self._residuals.append(residuals)
 
-        self._residuals.append(residuals)
+            # compute the sigma-clipped mean along the 3D stack
+            with warnings.catch_warnings():
+                warnings.simplefilter('ignore', category=RuntimeWarning)
+                warnings.simplefilter('ignore', category=AstropyUserWarning)
+                residuals = self.sigclip(residuals, axis=0, masked=False,
+                                         return_bounds=False)
+                if HAS_BOTTLENECK:
+                    residuals = bottleneck.nanmean(residuals, axis=0)
+                else:
+                    residuals = np.nanmean(residuals, axis=0)
 
-        # compute the sigma-clipped median along the 3D stack
-        with warnings.catch_warnings():
-            warnings.simplefilter('ignore', category=RuntimeWarning)
-            warnings.simplefilter('ignore', category=AstropyUserWarning)
-            residuals = self.sigclip(residuals, axis=0, masked=False,
-                                     return_bounds=False)
-            if HAS_BOTTLENECK:
-                residuals = bottleneck.nanmedian(residuals, axis=0)
-            else:
-                residuals = np.nanmedian(residuals, axis=0)
+            self._residuals_sigclip.append(residuals)
 
-        self._residuals_sigclip.append(residuals)
+            # interpolate any missing data (np.nan)
+            mask = ~np.isfinite(residuals)
+            if np.any(mask):
+                residuals = _interpolate_missing_data(residuals, mask,
+                                                      method='cubic')
 
-        # interpolate any missing data (np.nan)
-        mask = ~np.isfinite(residuals)
-        if np.any(mask):
-            residuals = _interpolate_missing_data(residuals, mask,
-                                                  method='cubic')
+                # fill any remaining nans (outer points) with zeros
+                residuals[~np.isfinite(residuals)] = 0.
 
-            # fill any remaining nans (outer points) with zeros
-            residuals[~np.isfinite(residuals)] = 0.
+            self._residuals_interp.append(residuals)
 
-        self._residuals_interp.append(residuals)
+            # add the residuals to the previous ePSF image
+            new_epsf = epsf._data + residuals
 
-        # add the residuals to the previous ePSF image
-        new_epsf = epsf.normalized_data + residuals
+            epsf = EPSFModel(data=new_epsf, origin=epsf.origin,
+                             oversampling=epsf.oversampling,
+                             norm_radius=epsf._norm_radius,
+                             shift_val=epsf._shift_val, normalize=False)
 
-        # smooth the ePSF
-        new_epsf = self._smooth_epsf(new_epsf)
+            # smooth and recenter the ePSF
+            epsf._data = self._smooth_epsf(epsf._data)
+            epsf._data = self._recenter_epsf(epsf, centroid_func=self.recentering_func)
 
-        # recenter the ePSF
-        new_epsf = self._recenter_epsf(new_epsf, epsf,
-                                       centroid_func=self.recentering_func,
-                                       box_size=self.recentering_boxsize,
-                                       maxiters=self.recentering_maxiters,
-                                       center_accuracy=1.0e-4)
+        # return the new ePSF object, but with undersampled grid pixel coordinates
+        xcenter = (epsf._data.shape[1] - 1) / 2. / epsf.oversampling[0]
+        ycenter = (epsf._data.shape[0] - 1) / 2. / epsf.oversampling[1]
 
-        # normalize the ePSF data
-        new_epsf /= np.sum(new_epsf, dtype=np.float64)
-
-        # return the new ePSF object
-        xcenter = (new_epsf.shape[1] - 1) / 2.
-        ycenter = (new_epsf.shape[0] - 1) / 2.
-
-        epsf_new = EPSFModel(data=new_epsf, origin=(xcenter, ycenter),
-                             normalize=False, oversampling=epsf.oversampling)
-
-        return epsf_new
+        return EPSFModel(data=epsf._data, origin=(xcenter, ycenter),
+                         oversampling=epsf.oversampling,
+                         norm_radius=epsf._norm_radius,
+                         shift_val=epsf._shift_val)
 
     def build_epsf(self, stars, init_epsf=None):
         """
@@ -775,16 +715,13 @@ class EPSFBuilder:
         """
 
         iter_num = 0
-        center_dist_sq = self.center_accuracy_sq + 1.
         centers = stars.cutout_center_flat
         n_stars = stars.n_stars
         fit_failed = np.zeros(n_stars, dtype=bool)
         epsf = init_epsf
         dt = 0.
 
-        while (iter_num < self.maxiters and
-               np.max(center_dist_sq) >= self.center_accuracy_sq and
-               not np.all(fit_failed)):
+        while (iter_num < self.maxiters and not np.all(fit_failed)):
 
             t_start = time.time()
             iter_num += 1
@@ -890,10 +827,10 @@ def _interpolate_missing_data(data, mask, method='cubic'):
     data_interp = np.array(data, copy=True)
 
     if len(data_interp.shape) != 2:
-        raise ValueError('data must be a 2D array.')
+        raise ValueError("'data' must be a 2D array.")
 
     if mask.shape != data.shape:
-        raise ValueError('mask and data must have the same shape.')
+        raise ValueError("'mask' and 'data' must have the same shape.")
 
     y, x = np.indices(data_interp.shape)
     xy = np.dstack((x[~mask].ravel(), y[~mask].ravel()))[0]