+++ photutils/psf/models.py
@@ -91,6 +91,10 @@ class FittableImageModel(Fittable2DModel):
         data) to the flux of model's data.
         Then, best fitted value of the `flux` model
         parameter will represent an aperture-corrected flux of the target star.
+        In the case of aperture correction, ``normalization_correction`` should
+        be a value larger than one, as the total flux, including regions outside
+        of the aperture, should be larger than the flux inside the aperture,
+        and thus the correction is applied as an inversely multiplied factor.
 
     fill_value : float, optional
         The value to be returned by the `evaluate` or
@@ -149,6 +153,15 @@ class FittableImageModel(Fittable2DModel):
         # set the origin of the coordinate system in image's pixel grid:
         self.origin = origin
 
+        flux = self._initial_norm(flux, normalize)
+
+        super().__init__(flux, x_0, y_0)
+
+        # initialize interpolator:
+        self.compute_interpolator(ikwargs)
+
+    def _initial_norm(self, flux, normalize):
+
         if flux is None:
             if self._img_norm is None:
                 self._img_norm = self._compute_raw_image_norm(self._data)
@@ -156,10 +169,7 @@ class FittableImageModel(Fittable2DModel):
 
         self._compute_normalization(normalize)
 
-        super().__init__(flux, x_0, y_0)
-
-        # initialize interpolator:
-        self.compute_interpolator(ikwargs)
+        return flux
 
     def _compute_raw_image_norm(self, data):
         """
@@ -487,17 +497,241 @@ class FittableImageModel(Fittable2DModel):
 
 class EPSFModel(FittableImageModel):
     """
-    A subclass of `FittableImageModel`. A fittable ePSF model.
+    A class that models an effective PSF (ePSF).
+
+    While this class is a subclass of `FittableImageModel`, it is very similar.
+    The primary differences/motivation are a few additional  parameters necesary
+    specifically for ePSFs.
+
+    Parameters
+    ----------
+    oversampling : int or tuple of two int, optional
+        The oversampling factor(s) of the model in the ``x`` and ``y`` directions.
+        If ``oversampling`` is a scalar it will be treated as being the same in both
+        x and y; otherwise a tuple of two floats will be treated as
+        ``(x_oversamp, y_oversamp)``.
+    norm_radius : float, optional
+        The radius inside which the ePSF is normalized by the sum over
+        undersampled integer pixel values inside a circular aperture.
+    shift_val : float, optional
+        The fractional undersampled pixel amount (equivalent to an integer
+        oversampled pixel value) at which to evaluate the asymmetric
+        ePSF centroid corrections.
     """
 
-    def __init__(self, data, flux=1.0, x_0=0, y_0=0, normalize=True,
-                 normalization_correction=1.0, origin=None, oversampling=1.,
-                 fill_value=0., ikwargs={}):
+    flux = Parameter(description='Intensity scaling factor for image data.',
+                     default=1.0)
+    x_0 = Parameter(description='X-position of a feature in the image in '
+                    'the output coordinate grid on which the model is '
+                    'evaluated.', default=0.0)
+    y_0 = Parameter(description='Y-position of a feature in the image in '
+                    'the output coordinate grid on which the model is '
+                    'evaluated.', default=0.0)
 
-        super().__init__(
-            data=data, flux=flux, x_0=x_0, y_0=y_0, normalize=normalize,
-            normalization_correction=normalization_correction, origin=origin,
-            oversampling=oversampling, fill_value=fill_value, ikwargs=ikwargs)
+    def __init__(self, data, flux=flux.default, x_0=x_0.default, y_0=y_0.default,
+                 origin=None, oversampling=1, fill_value=0.0, norm_radius=5.5,
+                 shift_val=0.5, normalize=True, normalization_correction=1.0,
+                 ikwargs={}):
+        self._norm_radius = norm_radius
+        self._shift_val = shift_val
+
+        super().__init__(data=data, flux=flux, x_0=x_0, y_0=y_0, normalize=normalize,
+                         normalization_correction=normalization_correction,
+                         origin=origin, oversampling=oversampling, fill_value=fill_value,
+                         ikwargs=ikwargs)
+
+    def _initial_norm(self, flux, normalize):
+        if flux is None:
+            if self._img_norm is None:
+                self._img_norm = self._compute_raw_image_norm(self._data,
+                                                              self._norm_radius)
+            flux = self._img_norm
+
+        if normalize:
+            self._compute_normalization()
+        else:
+            self._img_norm = self._compute_raw_image_norm(self._data,
+                                                          self._norm_radius)
+
+    def _compute_raw_image_norm(self, data, radius):
+        """
+        Helper function that computes the normalization of input image data.
+        This quantity is computed as the sum of all undersampled integer pixel
+        values within radius pixels of the center of the ePSF.
+        """
+
+        # First need the indices of each axis at the oversampled resolution;
+        # if oversampling = 4 then x = [0, 0.25, 0.5, 0.75, ...]
+        x = np.arange(self._nx, dtype=np.float64) / self.oversampling[0]
+        y = np.arange(self._ny, dtype=np.float64) / self.oversampling[1]
+        # Take indices where the undersampled grid is an integer -- i.e., the
+        # actual undersampled grid -- and find the cut where
+        # sqrt(dx**2 + dy**2) <= radius
+        x_0, y_0 = int((self._nx - 1) / 2), int((self._ny - 1) / 2)
+        # However, as we are in units of the undersampled grid, we must convert
+        # to undersampled units by the same factor of oversampling
+        x_0 /= self.oversampling[0]
+        y_0 /= self.oversampling[1]
+        # When checking if the index is at the center of a pixel, we check such
+        # that the index number is half that of the oversampling -- if we
+        # oversample by a factor 4 then the middle pixel of the 0th large pixel
+        # is 2 ([0, 1, 2, 3, 4]). For this to work we require oversampling to be
+        # an even number; otherwise, the ``middle'' pixel will be halfway between
+        # two oversampled pixels.
+        over_index_middle = 1 / 2
+        cut = (((x.reshape(1, -1) - x_0)**2 + (y.reshape(-1, 1) - y_0)**2 <=
+                radius**2) & (x.reshape(1, -1) % 1.0 == over_index_middle) &
+               (y.reshape(-1, 1) % 1.0 == over_index_middle))
+        data = self._data
+
+        return np.sum(data[cut], dtype=np.float64)
+
+    def _compute_normalization(self):
+        """
+        Helper function that computes (corrected) normalization factor
+        of the original image data. For the ePSF this is defined as the
+        sum over the inner N (default=5.5) pixels of the non-oversampled
+        image. Will re-normalize the data to the value calculated.
+        """
+
+        if self._img_norm is None:
+            if np.sum(self._data) == 0:
+                self._img_norm = 1
+            else:
+                self._img_norm = self._compute_raw_image_norm(self._data,
+                                                              self._norm_radius)
+
+        if self._img_norm != 0.0 and np.isfinite(self._img_norm):
+            self._data /= (self._img_norm * self._normalization_correction)
+            self._normalization_status = 0
+        else:
+            self._normalization_status = 1
+            self._img_norm = 1
+            warnings.warn("Overflow encountered while computing "
+                          "normalization constant. Normalization "
+                          "constant will be set to 1.", NonNormalizable)
+
+    def _set_oversampling(self, value):
+        try:
+            value = np.atleast_1d(value).astype(int)
+            if len(value) == 1:
+                value = np.repeat(value, 2)
+            # We need oversampling to be a factor of 2 for ``middle of pixel''
+            # in the undersampled regime to have a pixel placed at it in the
+            # oversampled regime.
+            if np.any(value % 2 != 0) and np.logical_not(np.all(value == 1)):
+                raise ValueError('Oversampling factor must be a multiple of two')
+        except ValueError:
+            raise ValueError('Oversampling factor must be a scalar')
+        if np.any(value <= 0):
+            raise ValueError('Oversampling factor must be greater than 0')
+
+        self._oversampling = value
+
+    def normalized_data(self):
+        """
+        Overloaded dummy function that also returns self._data, as the
+        normalization occurs within _compute_normalization in EPSFModel,
+        and as such self._data will sum, accounting for under/oversampled
+        pixels, to 1/self._normalization_correction. """
+        return self._data
+
+    @FittableImageModel.origin.setter
+    def origin(self, origin):
+        if origin is None:
+            self._x_origin = (self._nx - 1) / 2.0 / self.oversampling[0]
+            self._y_origin = (self._ny - 1) / 2.0 / self.oversampling[1]
+        elif (hasattr(origin, '__iter__') and len(origin) == 2):
+            self._x_origin, self._y_origin = origin
+        else:
+            raise TypeError("Parameter 'origin' must be either None or an "
+                            "iterable with two elements.")
+
+    def compute_interpolator(self, ikwargs={}):
+        """
+        Compute/define the interpolating spline. This function can be overriden
+        in a subclass to define custom interpolators.
+
+        Parameters
+        ----------
+        ikwargs : dict, optional
+
+            Additional optional keyword arguments. Possible values are:
+
+            - **degree** : int, tuple, optional
+                Degree of the interpolating spline. A tuple can be used to
+                provide different degrees for the X- and Y-axes.
+                Default value is degree=3.
+
+            - **s** : float, optional
+                Non-negative smoothing factor. Default value s=0 corresponds to
+                interpolation.
+                See :py:class:`~scipy.interpolate.RectBivariateSpline` for more
+                details.
+
+        Notes
+        -----
+            * When subclassing :py:class:`FittableImageModel` for the
+              purpose of overriding :py:func:`compute_interpolator`,
+              the :py:func:`evaluate` may need to overriden as well depending
+              on the behavior of the new interpolator. In addition, for
+              improved future compatibility, make sure
+              that the overriding method stores keyword arguments ``ikwargs``
+              by calling ``_store_interpolator_kwargs`` method.
+            * Use caution when modifying interpolator's degree or smoothness in
+              a computationally intensive part of the code as it may decrease
+              code performance due to the need to recompute interpolator.
+        """
+        from scipy.interpolate import RectBivariateSpline
+
+        if 'degree' in ikwargs:
+            degree = ikwargs['degree']
+            if hasattr(degree, '__iter__') and len(degree) == 2:
+                degx = int(degree[0])
+                degy = int(degree[1])
+            else:
+                degx = int(degree)
+                degy = int(degree)
+            if degx < 0 or degy < 0:
+                raise ValueError("Interpolator degree must be a non-negative "
+                                 "integer")
+        else:
+            degx = 3
+            degy = 3
+
+        if 's' in ikwargs:
+            smoothness = ikwargs['s']
+        else:
+            smoothness = 0
+
+        # Interpolator must be set to interpolate on the undersampled pixel
+        # grid, going from 0 to len(undersampled_grid)
+        x = np.arange(self._nx, dtype=np.float) / self.oversampling[0]
+        y = np.arange(self._ny, dtype=np.float) / self.oversampling[1]
+        self.interpolator = RectBivariateSpline(
+            x, y, self._data.T, kx=degx, ky=degy, s=smoothness)
+
+        self._store_interpolator_kwargs(ikwargs)
+
+    def evaluate(self, x, y, flux, x_0, y_0):
+        """
+        Evaluate the model on some input variables and provided model
+        parameters.
+        """
+
+        xi = np.asarray(x) - x_0 + self._x_origin
+        yi = np.asarray(y) - y_0 + self._y_origin
+
+        evaluated_model = flux * self.interpolator.ev(xi, yi)
+
+        if self._fill_value is not None:
+            # find indices of pixels that are outside the input pixel grid and
+            # set these pixels to the 'fill_value':
+            invalid = (((xi < 0) | (xi > (self._nx - 1) / self.oversampling[0])) |
+                       ((yi < 0) | (yi > (self._ny - 1) / self.oversampling[1])))
+            evaluated_model[invalid] = self._fill_value
+
+        return evaluated_model
 
 
 class GriddedPSFModel(Fittable2DModel):
@@ -759,11 +993,11 @@ class IntegratedGaussianPRF(Fittable2DModel):
     ----------
     sigma : float
         Width of the Gaussian PSF.
-    flux : float (default 1)
+    flux : float, optional
         Total integrated flux over the entire PSF
-    x_0 : float (default 0)
+    x_0 : float, optional
         Position of the peak in x direction.
-    y_0 : float (default 0)
+    y_0 : float, optional
         Position of the peak in y direction.
 
     Notes