+++ photutils/utils/misc.py
@@ -4,10 +4,26 @@ This module provides tools to return the installed astropy and photutils
 versions.
 """
 
+from astropy.utils import deprecated
+
 __all__ = ['get_version_info']
 
 
+@deprecated('0.7')
 def get_version_info():
+    """
+    Return astropy and photutils versions.
+
+    Returns
+    -------
+    result : str
+        The astropy and photutils versions.
+    """
+
+    return _get_version_info()
+
+
+def _get_version_info():
     """
     Return astropy and photutils versions.